Some progress on getting parasail to run

+++ FindSimSeqs/Alignment.h
@@ -7,6 +7,6 @@ void fndMtchngCrds(const char* cigStr, const int32_t &strLen, uint32_t &start, u
 
 //This function parses a parasail cigar and counts the number of leading and trailing deletion, the length of the alignment in bet-
 //ween and the total number of matched bases
-void prsCgr(const parasail_cigar_t &cgr, uint32_t &start, uint32_t &end, uint32_t &aLen, uint32_t &nMtchs);
+void prsCgr(const parasail_cigar_t &cgr, uint32_t &start, uint32_t &end, float &aLen, float &nMtchs);
 
 #endif
